http-req: split up into multiple files
